LayoutTests/fast/page-color-sampling/color-sampling-ignores-subscrollers-expected.txt

PASS colors.top is null
PASS colors.left is null
PASS colors.right is null
PASS colors.bottom is "rgb(225, 225, 225)"
PASS successfullyParsed is true

TEST COMPLETE
